fix(test): ensure that binary is built before the test runs
The `app.rs` test requires that the `binsider` binary has been built. By
moving the test into the `tests` directory, `cargo` will ensure that the
binary has been built before the test can be run.
